requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.49.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.0
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
mplfinance==0.12.10b0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
TA-Lib @ file:///C:/D/Ryan/TAMU_class/STAT654/Project/TA_Lib-0.4.24-cp38-cp38-win_amd64.whl#sha256=26d934a971a39a18e265f8b5999eb1a7d6c8fb8fca36a88881fcbb05d6138bc4
threadpoolctl==3.3.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.18.0
